backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv
python-multipart
pydantic
matplotlib
